backend/requirements.txt
Flask==2.2.0
flask-restx==0.5.1
Werkzeug==2.2.0
gunicorn==20.1.0
mysql-connector
pymysql
PyJWT==2.1.0
sqlalchemy==2.0.9
Flask-Cors==3.0.10
marshmallow==3.19.0
bcrypt==4.0.1